LPD-90495 moves JiraAssetObject to model package
